Crate libre_pvz_resources

librePvZ-resources: resource loading logics for librePvZ.

Modules

animation

Sprite and animation API.

bevy

Bevy related utilities.

cached

Serialize and deserialize as some kind of raw form (a String or a PathBuf usually), used as a key indexing into some collection (a boxed slice, or the asset collection). Keep also a cached handle to speed up subsequent access (an integer index, or a Handle).

model

Models incorporating animations.