Module libre_pvz_resources::bevy

Bevy related utilities.

Re-exports

pub use loader::AddTwoStageAsset;

pub use loader::TwoStageAssetPlugin;

pub use animation::Animation;

Modules

animation

Animation related support using Bevy and libre_pvz_animation.

loader

Loaders for bincode, JSON, and YAML files. These files can decode into different data structures, and therefore require customisation over bevy’s AssetLoader. We do so by requiring an additional “secondary extension” in asset file names. For example, a file named “Peashooter.anim.bin” is treated as encoded in bincode, and has a resource type “anim”.

Structs

ResourcesPlugin

Resources plugin.