Module libre_pvz_resources::bevy::loader

Loaders for bincode, JSON, and YAML files. These files can decode into different data structures, and therefore require customisation over bevy’s AssetLoader. We do so by requiring an additional “secondary extension” in asset file names. For example, a file named “Peashooter.anim.bin” is treated as encoded in bincode, and has a resource type “anim”.

Structs

LoaderRegistry

Registry for asset loaders.

TwoStageAssetPlugin

Plugin for two-stage asset loading.

Traits

AddTwoStageAsset

Extension to the App API for adding two-stage asset loaders.

TwoStageAssetLoader

A new trait for two-stage asset loading, in place of AssetLoader:

Functions

default_post_process

Default post processing: no processing at all, submit the decoded data as loaded asset.