Trait libre_pvz_resources::bevy::loader::TwoStageAssetLoader

pub trait TwoStageAssetLoader: Send + Sync + 'static {
    type Repr: Decode + DeserializeOwned;

    fn extension(&self) -> &str;
    fn post_process(
        &self, 
        repr: Self::Repr, 
        load_context: &mut LoadContext<'_>
    ) -> Result<()>;
}

A new trait for two-stage asset loading, in place of AssetLoader:

Decode from bincode/JSON/YAML etc. to get structural data

Post-processing the structural data, transforming & adding dependencies This is meant to support loading the same data structures stored in different serialised forms, and share their post-processing logic.

Required Associated Types

type Repr: Decode + DeserializeOwned

The decoded representation for this asset.

Required Methods

fn extension(&self) -> &str

The “secondary file extension” this asset is associated to.

fn post_process(
&self,
repr: Self::Repr,
load_context: &mut LoadContext<'_>
) -> Result<()>

The post-processing logic: transform the Repr to a more compact in-memory form, require loading the dependencies and store their handles in the appropriate locations, and submit the resulting asset to the asset loader.