Module libre_pvz_resources::cached

Serialize and deserialize as some kind of raw form (a String or a PathBuf usually), used as a key indexing into some collection (a boxed slice, or the asset collection). Keep also a cached handle to speed up subsequent access (an integer index, or a Handle).

Structs

Cached

Raw key storage with cached handle.

SortedSlice

Raw key storage with cached handle.

Traits

ContainerWithKey

Container of key-value pairs, equipped with a handle type to speed up lookup.

EntryWithKey

Entries in linear homogeneous collections.