Module libre_pvz_resources::dynamic

Dynamic resource support. For simplicity, we use shared global state for registry.

Structs

DynamicRegistry

Type registry for dynamic content (de)serialization.

DynamicResource

Resource, but without a statically-known type.

ReflectAnyResource

TypeData providing support for AnyResource trait.

Traits

AnyResource

Resource types with dynamic (de)serialization. Serialization through bincode is fixed to using the standard configuration, and cannot be changed at runtime, due to the current design of bincode API.

ErasedResource

Allow dynamic (de)serialization for trait objects.