Trait libre_pvz_resources::dynamic::AnyResource

pub trait AnyResource: Reflect + ErasedSerialize + Send + Sync + 'static {
    // Required methods
    fn as_reflect(&self) -> &dyn Reflect;
    fn as_reflect_mut(&mut self) -> &mut dyn Reflect;
    fn erased_deserialize(
        src: &mut dyn ErasedDeserializer<'_>
    ) -> Result<Box<dyn AnyResource>, Error>
       where Self: Sized;
    fn erased_encode(&self, writer: &mut dyn Writer) -> Result<(), EncodeError>;
    fn erased_decode(
        reader: &mut dyn Reader
    ) -> Result<Box<dyn AnyResource>, DecodeError>
       where Self: Sized;
}

Resource types with dynamic (de)serialization. Serialization through bincode is fixed to using the standard configuration, and cannot be changed at runtime, due to the current design of bincode API.

Required Methods

fn as_reflect(&self) -> &dyn Reflect

Convert to a Reflect trait object.

fn as_reflect_mut(&mut self) -> &mut dyn Reflect

Convert to a mutable Reflect trait object.

fn erased_deserialize( src: &mut dyn ErasedDeserializer<'_> ) -> Result<Box<dyn AnyResource>, Error>
where Self: Sized,

Deserialize from an erased deserializer.

fn erased_encode(&self, writer: &mut dyn Writer) -> Result<(), EncodeError>

Encode as bincode to a given Writer using standard configuration.

fn erased_decode( reader: &mut dyn Reader ) -> Result<Box<dyn AnyResource>, DecodeError>
where Self: Sized,

Decode as bincode from a given Reader using standard configuration.