Crate libre_pvz_resources

librePvZ-resources: resource loading logics for librePvZ.

Re-exports

pub use once_cell;

Modules

animation

Sprite and animation API.

cached

Serialize and deserialize as some kind of raw form (a String or a PathBuf usually), used as a key indexing into some collection (a boxed slice, or the asset collection). Keep also a cached handle to speed up subsequent access (an integer index, or a Handle).

dynamic

Dynamic resource support. For simplicity, we use shared global state for registry.

loader

Loaders for bincode, JSON, and YAML files. These files can decode into different data structures, and therefore require customisation over bevy’s AssetLoader. We do so by requiring an additional “secondary extension” in asset file names. For example, a file named “Peashooter.anim.bin” is treated as encoded in bincode, and has a resource type “anim”.

model

Models incorporating animations.

Macros

asset_ext

Generate proper values for AssetExtensions. Always prefer this macro instead of manually constructing AssetExtensions values so that the program behaviour is consistent.

cache_known_states

Generate StateIndex and StateIndex::cache for cached known states.

mark_trait_as_dynamic_resource

Mark a trait available for dynamic (de)serialization.

Structs

ResourcesPlugins

Resources plugin group.