Module libre_pvz_resources::model

Models incorporating animations.

Structs

Attachment

Attachment, useful for separating different movable parts in a single entity.

AutoNullTrigger

Automatically apply the None trigger for the respective model.

CoolDown

Cool down component for state transitions.

MarkerRegistry

Registry for marker components.

Model

Model: animation together with its association.

ModelPlugin

Model plugin.

ModelState

Keep track of the current state of the model.

PlantMeta

Plant meta information.

State

State controls the appearance and behaviours.

StateTransition

Transition from one state to another.

StateTransitionEvent

ModelState transition events.

TransitionTrigger

Request to trigger a ModelState transition.

Enums

ModelSystem

Labels for model-related systems.

Traits

MarkerRegistryExt

Extend the App for registering marker components.